macOS: refactor directory structure
