+++ macos/Classes/FlutterAcrylicPlugin.swift
@@ -1,480 +1,6 @@
 import Cocoa
 import FlutterMacOS
 
-public class BlurryContainerViewController: NSViewController {
-    public let flutterViewController = FlutterViewController()
-
-    public init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-
-    required public init?(coder: NSCoder) {
-        fatalError()
-    }
-
-    override public func loadView() {
-        let blurView = NSVisualEffectView()
-        blurView.autoresizingMask = [.width, .height]
-        blurView.blendingMode = .behindWindow
-        blurView.state = .followsWindowActiveState
-        if #available(macOS 10.14, *) {
-            blurView.material = .underWindowBackground
-        }
-        self.view = blurView
-    }
-
-    override public func viewDidLoad() {
-        super.viewDidLoad()
-
-        self.addChild(flutterViewController)
-
-        flutterViewController.view.frame = self.view.bounds
-        flutterViewController.view.autoresizingMask = [.width, .height]
-        self.view.addSubview(flutterViewController.view)
-    }
-}
-
-private class EffectIDToMaterialConverter {
-    @available(macOS 10.14, *)
-    public static func getMaterialFromEffectID(effectID: NSNumber) -> NSVisualEffectView.Material {
-        switch(effectID) {
-        /* Try to mimic the behavior of the following effects as
-           closely as possible: */
-        case 0: // disabled
-            return NSVisualEffectView.Material.windowBackground
-            
-        case 1: // solid
-            return NSVisualEffectView.Material.windowBackground
-            
-        case 2: // transparent
-            return NSVisualEffectView.Material.underWindowBackground
-            
-        case 3: // aero
-            return NSVisualEffectView.Material.hudWindow
-            
-        case 4: // acrylic
-            return NSVisualEffectView.Material.fullScreenUI
-            
-        case 5: // mica
-            return NSVisualEffectView.Material.headerView
-
-        case 6: // tabbed
-            return NSVisualEffectView.Material.headerView
-            
-        case 7: // titlebar
-            return NSVisualEffectView.Material.titlebar
-                    
-        case 8: // selection
-            return NSVisualEffectView.Material.selection
-            
-        case 9: // menu
-            return NSVisualEffectView.Material.menu
-            
-        case 10: // popover
-            return NSVisualEffectView.Material.popover
-            
-        case 11: // sidebar
-            return NSVisualEffectView.Material.sidebar
-            
-        case 12: // headerView
-            return NSVisualEffectView.Material.headerView
-            
-        case 13: // sheet
-            return NSVisualEffectView.Material.sheet
-            
-        case 14: // windowBackground
-            return NSVisualEffectView.Material.windowBackground
-            
-        case 15: // hudWindow
-            return NSVisualEffectView.Material.hudWindow
-            
-        case 16: // fullScreenUI
-            return NSVisualEffectView.Material.fullScreenUI
-            
-        case 17: // toolTip
-            return NSVisualEffectView.Material.toolTip
-            
-        case 18: // contentBackground
-            return NSVisualEffectView.Material.contentBackground
-            
-        case 19: // underWindowBackground
-            return NSVisualEffectView.Material.underWindowBackground
-            
-        case 20: // underPageBackground
-            return NSVisualEffectView.Material.underPageBackground
-            
-        default:
-            return NSVisualEffectView.Material.windowBackground
-        }
-    }
-}
-
-public class MainFlutterWindowManipulator {
-    private static var mainFlutterWindow: NSWindow?
-    
-    private static func printNotStartedWarning() {
-        print("Warning: The MainFlutterWindowManipulator has not been started. Please make sure the flutter_acrylic plugin is initialized correctly in your MainFlutterWindow.swift file.")
-    }
-    
-    public static func start(mainFlutterWindow: NSWindow) {
-        self.mainFlutterWindow = mainFlutterWindow
-        
-        showTitle()
-        makeTitlebarOpaque()
-        disableFullSizeContentView()
-        setWindowBackgroundColorToDefaultColor()
-    }
-    
-    public static func hideTitle() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.titleVisibility = NSWindow.TitleVisibility.hidden
-    }
-    
-    public static func showTitle() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.titleVisibility = NSWindow.TitleVisibility.visible
-    }
-    
-    public static func makeTitlebarTransparent() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.titlebarAppearsTransparent = true
-    }
-    
-    public static func makeTitlebarOpaque() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.titlebarAppearsTransparent = false
-    }
-    
-    public static func enableFullSizeContentView() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.styleMask.insert(.fullSizeContentView)
-    }
-    
-    public static func disableFullSizeContentView() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.styleMask.remove(.fullSizeContentView)
-    }
-    
-    public static func zoomWindow() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.setIsZoomed(true)
-    }
-    
-    public static func unzoomWindow() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.setIsZoomed(false)
-    }
-    
-    public static func isWindowZoomed() -> Bool {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return false
-        }
-        
-        return self.mainFlutterWindow!.isZoomed
-    }
-    
-    public static func enterFullscreen() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        if (!isWindowFullscreened()) {
-            self.mainFlutterWindow!.toggleFullScreen(self)
-        }
-    }
-    
-    public static func exitFullscreen() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        if (isWindowFullscreened()) {
-            self.mainFlutterWindow!.toggleFullScreen(self)
-        }
-    }
-    
-    public static func isWindowFullscreened() -> Bool {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return false
-        }
-        
-        let isFullscreenEnabled = self.mainFlutterWindow!.styleMask.contains(NSWindow.StyleMask.fullScreen)
-        return isFullscreenEnabled
-    }
-    
-    public static func hideZoomButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.zoomButton)!.isHidden = true
-    }
-    
-    public static func showZoomButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.zoomButton)!.isHidden = false
-    }
-    
-    public static func hideMiniaturizeButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.miniaturizeButton)!.isHidden = true
-    }
-    
-    public static func showMiniaturizeButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.miniaturizeButton)!.isHidden = false
-    }
-    
-    public static func hideCloseButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.closeButton)!.isHidden = true
-    }
-    
-    public static func showCloseButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.closeButton)!.isHidden = false
-    }
-    
-    public static func enableZoomButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.zoomButton)!.isEnabled = true
-    }
-
-    public static func disableZoomButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.zoomButton)!.isEnabled = false
-    }
-
-    public static func enableMiniaturizeButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.miniaturizeButton)!.isEnabled = true
-    }
-
-    public static func disableMiniaturizeButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.miniaturizeButton)!.isEnabled = false
-    }
-
-    public static func enableCloseButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.closeButton)!.isEnabled = true
-    }
-
-    public static func disableCloseButton() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.standardWindowButton(.closeButton)!.isEnabled = false
-    }
-    
-    public static func setDocumentEdited() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.isDocumentEdited = true
-    }
-    
-    public static func setDocumentUnedited() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.isDocumentEdited = false
-    }
-    
-    public static func setRepresentedFilename(filename: String) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.representedFilename = filename
-    }
-    
-    public static func setRepresentedURL(url: String) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.representedURL = URL(string: url)
-    }
-    
-    public static func isWindowInLiveResize() -> Bool {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return false
-        }
-        
-        return self.mainFlutterWindow!.inLiveResize
-    }
-    
-    public static func setWindowAlphaValue(alphaValue: CGFloat) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.alphaValue = alphaValue
-    }
-    
-    public static func isWindowVisible() -> Bool {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return false
-        }
-        
-        return self.mainFlutterWindow!.isVisible
-    }
-    
-    public static func setWindowBackgroundColorToDefaultColor() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.backgroundColor = .windowBackgroundColor
-    }
-    
-    public static func setWindowBackgroundColorToClear() {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow!.backgroundColor = .clear
-    }
-    
-    public static func setBlurViewState(state: NSVisualEffectView.State) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        let blurryContainerViewController = self.mainFlutterWindow?.contentViewController as! BlurryContainerViewController;
-        (blurryContainerViewController.view as! NSVisualEffectView).state = state
-    }
-    
-    @available(macOS 10.14, *)
-    public static func setAppearance(dark: Bool) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        self.mainFlutterWindow?.contentView?.superview?.appearance = NSAppearance(named: dark ? .darkAqua : .aqua)
-        
-        self.mainFlutterWindow?.invalidateShadow()
-    }
-    
-    @available(macOS 10.14, *)
-    public static func setMaterial(material: NSVisualEffectView.Material) {
-        if (self.mainFlutterWindow == nil) {
-            printNotStartedWarning()
-            return
-        }
-        
-        let blurryContainerViewController = self.mainFlutterWindow?.contentViewController as! BlurryContainerViewController;
-        (blurryContainerViewController.view as! NSVisualEffectView).material = material
-        
-        self.mainFlutterWindow?.invalidateShadow()
-    }
-    
-    @available(macOS 10.14, *)
-    public static func setEffect(material: NSVisualEffectView.Material) {
-        setMaterial(material: material)
-    }
-    
-    public static func getTitlebarHeight() -> CGFloat {
-        let windowFrameHeight = (self.mainFlutterWindow!.contentView?.frame.height)!
-        let contentLayoutRectHeight = self.mainFlutterWindow!.contentLayoutRect.height
-        let fullSizeContentViewNoContentAreaHeight = windowFrameHeight - contentLayoutRectHeight
-        return fullSizeContentViewNoContentAreaHeight
-    }
-}
-
 public class FlutterAcrylicPlugin: NSObject, FlutterPlugin {
     private var registrar: FlutterPluginRegistrar!;
     private var channel: FlutterMethodChannel!
